Fixed issue with editor not letting content be edited
The problem was that the content was being refreshed on each update to the form, so no edits were being maintained. The updateFormData and clearForm methods need to be called when the editor changed visibility and at no other time, so a watch function is more suitable than trying to bind these methods to a lifecycle hook.